Rename listOrders method to getAllOrders
The `listOrders` method in `ShopService` has been renamed to `getAllOrders`. This change was made to make the method name more intuitive. Additionally, a refactor was made in `Order` class to implement a Lombok `@With` annotation in the `OrderStatus` field. This allows the model to perform a copy of an object with a changed field more efficiently. In `OrderRepository`, instead of creating a new Order object when updating the order status, we are now using the `withOrderStatus` method, which reduces the code complexity.