Added order creation & retrieving oldest order per status feature
The main method in `Main.java` was updated to import various modules and support the creation of orders. An `Order` class, `OrderRepository`, `ProductRepository`, `IdService`, and `ShopService` were all instantiated. In addition, three instances of `Order` were created and added to the `ShopService`.

In `ShopService.java`, a new method `getOldestOrderPerStatus` was added. This method retrieves the oldest order for each `OrderStatus`.

Also, `OrderRepository.java` was updated. The initialization of `orderMap` was moved from the constructor to the class field declaration. All orders can now be retrieved using the `getAllOrders` method instead of `listOrders`.